simplifying cli.py
