+++ tools/release/cli_tools/commands/version.py
@@ -1,0 +1,158 @@
+import argparse
+import shutil
+from pathlib import Path
+
+from tools.release.changelog.release_workflow import DEFAULT_CHANGELOG_SCRATCH_DIR
+from tools.release.version.models import Version
+from tools.release.version.sync import apply_version_update, resolve_debian_revision
+from tools.release.version.validate import (
+    build_sync_required_message,
+    get_release_state,
+    require_release_files,
+    require_synced_release_state,
+)
+from tools.release.cli_tools.print import print_release_state, print_planned_changes
+
+def bump_version(version: Version, part: str) -> Version:
+    if part == "major":
+        return version.bump_major()
+    if part == "minor":
+        return version.bump_minor()
+    if part == "patch":
+        return version.bump_patch()
+    raise ValueError(f"Unsupported bump part: {part}")
+
+
+def cmd_check(args: argparse.Namespace) -> int:
+    state = get_release_state(args.repo_root)
+    print_release_state(state)
+
+    if state.has_structural_errors:
+        return 1
+    if state.has_drift:
+        print()
+        print(build_sync_required_message(state))
+        return 1
+    return 0
+
+
+def cmd_sync(args: argparse.Namespace) -> int:
+    repo_root = Path(args.repo_root).resolve()
+    state = require_release_files(repo_root)
+    paths = state.paths
+
+    canonical = state.versions.canonical
+    if canonical is None:
+        raise ValueError("Canonical VERSION could not be resolved")
+
+    debian_revision = resolve_debian_revision(
+        explicit_revision=args.debian_revision,
+        current_debian=state.versions.debian,
+    )
+
+    print(f"Canonical version: {canonical}")
+    print(f"Debian revision:   {debian_revision}")
+    print()
+
+    print_planned_changes(
+        current_canonical=state.versions.canonical,
+        current_meson=state.versions.meson,
+        current_package_json=state.versions.package_json,
+        current_debian=state.versions.debian,
+        target_version=canonical,
+        target_debian_revision=debian_revision,
+    )
+
+    if args.dry_run:
+        print("\nDry run only. No files were modified.")
+        return 0
+
+    should_clear_scratch = (
+        state.versions.debian is not None and state.versions.debian.upstream != canonical
+    )
+    apply_version_update(
+        paths=paths,
+        version=canonical,
+        debian_revision=debian_revision,
+        write_canonical=False,
+    )
+    if should_clear_scratch:
+        clear_changelog_scratch(repo_root, reason=f"VERSION transition to {canonical} via sync")
+    # Enforce that sync leaves the repo in a fully consistent release state.
+    _ = require_synced_release_state(repo_root)
+
+    print("\nSync complete.")
+    return 0
+
+
+def cmd_set_version(args: argparse.Namespace) -> int:
+    repo_root = Path(args.repo_root).resolve()
+    _ = require_synced_release_state(repo_root)
+
+    target_version = Version.parse(args.version)
+    debian_revision = args.debian_revision
+    state = get_release_state(repo_root)
+    paths = state.paths
+
+    print(f"Target version:    {target_version}")
+    print(f"Debian revision:   {debian_revision}")
+    print()
+
+    print_planned_changes(
+        current_canonical=state.versions.canonical,
+        current_meson=state.versions.meson,
+        current_package_json=state.versions.package_json,
+        current_debian=state.versions.debian,
+        target_version=target_version,
+        target_debian_revision=debian_revision,
+    )
+
+    if args.dry_run:
+        print("\nDry run only. No files were modified.")
+        return 0
+
+    previous_canonical = state.versions.canonical
+    apply_version_update(
+        paths=paths,
+        version=target_version,
+        debian_revision=debian_revision,
+        write_canonical=True,
+    )
+    if previous_canonical is not None and previous_canonical != target_version:
+        clear_changelog_scratch(repo_root, reason=f"VERSION bump {previous_canonical} -> {target_version}")
+
+    print("\nVersion update complete.")
+    return 0
+
+
+def cmd_bump(args: argparse.Namespace) -> int:
+    repo_root = Path(args.repo_root).resolve()
+    state = require_synced_release_state(repo_root)
+
+    current_version = state.versions.canonical
+    if current_version is None:
+        raise ValueError("Could not determine canonical version from VERSION")
+
+    target_version = bump_version(current_version, args.part)
+
+    print(f"Current version:   {current_version}")
+    print(f"Bump part:         {args.part}")
+    print(f"Target version:    {target_version}")
+    print()
+
+    set_args = argparse.Namespace(
+        repo_root=str(repo_root),
+        version=str(target_version),
+        debian_revision=args.debian_revision,
+        dry_run=args.dry_run,
+    )
+    return cmd_set_version(set_args)
+
+def clear_changelog_scratch(repo_root: Path, *, reason: str) -> None:
+    scratch_dir = (repo_root / DEFAULT_CHANGELOG_SCRATCH_DIR).resolve()
+    if not scratch_dir.exists():
+        return
+    if not scratch_dir.is_dir():
+        raise ValueError(f"Expected changelog scratch path to be a directory: {scratch_dir}")
+    shutil.rmtree(scratch_dir)
+    print(f"Cleared changelog scratch artifacts at {scratch_dir} ({reason}).")
